requirements.txt
langchain 
langchain_google_genai 
langchain_core 
pandas 
fastapi
setuptools
python-dotenv 
pyyaml
python-box
-e .